CLI: split context git history subtree
