+++ cmd/prescribe/cmds/context/git/history/set.go
@@ -1,88 +1,14 @@
-package git
+package history
 
 import (
 	"fmt"
 	"strconv"
 
 	"github.com/go-go-golems/prescribe/cmd/prescribe/cmds/helpers"
-	"github.com/go-go-golems/prescribe/internal/domain"
 	"github.com/spf13/cobra"
 )
 
-func effectiveGitHistoryConfig(data *domain.PRData) (domain.GitHistoryConfig, bool) {
-	if data != nil && data.GitHistory != nil {
-		return *data.GitHistory, true
-	}
-	return domain.DefaultGitHistoryConfig(), false
-}
-
-func newGitHistoryShowCmd() *cobra.Command {
-	return &cobra.Command{
-		Use:   "show",
-		Short: "Show effective git history config",
-		RunE: func(cmd *cobra.Command, args []string) error {
-			ctrl, err := helpers.NewInitializedController(cmd)
-			if err != nil {
-				return err
-			}
-			helpers.LoadDefaultSessionIfExists(ctrl)
-
-			data := ctrl.GetData()
-			cfg, explicit := effectiveGitHistoryConfig(data)
-			rangeSpec := fmt.Sprintf("%s..%s", data.TargetBranch, data.SourceBranch)
-
-			src := "defaults (missing in session.yaml)"
-			if explicit {
-				src = "session.yaml"
-			}
-
-			fmt.Printf("Git history config (%s)\n", src)
-			fmt.Printf("  enabled: %v\n", cfg.Enabled)
-			fmt.Printf("  max_commits: %d\n", cfg.MaxCommits)
-			fmt.Printf("  include_merges: %v\n", cfg.IncludeMerges)
-			fmt.Printf("  first_parent: %v\n", cfg.FirstParent)
-			fmt.Printf("  include_numstat: %v\n", cfg.IncludeNumstat)
-			fmt.Printf("  range: %s\n", rangeSpec)
-			return nil
-		},
-	}
-}
-
-func newGitHistoryEnableCmd(enable bool) *cobra.Command {
-	verb := "enable"
-	short := "Enable derived git history"
-	if !enable {
-		verb = "disable"
-		short = "Disable derived git history"
-	}
-
-	return &cobra.Command{
-		Use:   verb,
-		Short: short,
-		RunE: func(cmd *cobra.Command, args []string) error {
-			ctrl, err := helpers.NewInitializedController(cmd)
-			if err != nil {
-				return err
-			}
-			helpers.LoadDefaultSessionIfExists(ctrl)
-
-			data := ctrl.GetData()
-			cfg, _ := effectiveGitHistoryConfig(data)
-			cfg.Enabled = enable
-			data.GitHistory = &cfg
-
-			savePath := ctrl.GetDefaultSessionPath()
-			if err := ctrl.SaveSession(savePath); err != nil {
-				return err
-			}
-			fmt.Printf("Git history %sd\n", verb)
-			fmt.Printf("Session saved: %s\n", savePath)
-			return nil
-		},
-	}
-}
-
-func newGitHistorySetCmd() *cobra.Command {
+func newSetCmd() *cobra.Command {
 	var (
 		enabledStr        string
 		maxCommits        int
